requirements.txt
docling==2.72.0
docling-core==2.63.0
langchain-core==1.2.16
langchain-openai==1.1.7
python-dotenv==1.2.1
pypdf==6.7.5